renamed src to rtabm (better module name) and created first tests
