+++ rtabm/plot_FIG1.py
